feat: add nostr repost agent
